+++ CotEditor/Sources/RegexTextField.swift
@@ -8,7 +8,7 @@
 //
 //  ---------------------------------------------------------------------------
 //
-//  © 2018-2020 1024jp
+//  © 2018-2022 1024jp
 //
 //  Licensed under the Apache License, Version 2.0 (the "License");
 //  you may not use this file except in compliance with the License.
@@ -95,6 +95,15 @@ final class RegexTextField: NSTextField {
         super.textDidChange(notification)
         
         self.invalidateFieldEditor()
+    }
+    
+    
+    /// string value was directly changed
+    override var objectValue: Any? {
+        
+        didSet {
+            self.invalidateFieldEditor()
+        }
     }